feat: implement dashboard header redesign with feature flag
