+++ packages/frontend/src/components/common/Dashboard/DashboardHeaderV2.module.css
@@ -1,0 +1,4 @@
+.header {
+    border-bottom: 1px solid var(--mantine-color-ldGray-2);
+    z-index: 1;
+}
