+++ packages/frontend/src/mantine8Theme.ts
@@ -146,7 +146,8 @@ export const getMantine8ThemeOverride = (
                     root: {
                         fontFamily: theme.fontFamily,
                         fontWeight: 500,
-                        borderRadius: theme.radius.md,
+                        // TODO
+                        // borderRadius: theme.radius.md,
                     },
                 }),
                 defaultProps: {
